Completed Section 3 Create a Feature Component
The goal of this section is to segregate and loosely-couple the implementation details of specific components from the parent because as the application evolves it will be more maintainable.  Logically, it makes sense for smaller sub-components to focus on more specific tasks. This is done so that the separate component is reusable and the parent class's responsibilities are limited.
The parent component has control over the child component by a property binding.
An @Input decorator is used to ensure availability of the hero property for binding by the external HeroesComponent.